Admin BE Update & Add Seeder

+++ smashclub/src/main/java/com/backendsyndicate/smashclub/admin/repo/AdminMenuCategoryRepo.java
@@ -4,6 +4,6 @@ import com.backendsyndicate.smashclub.admin.model.AdminMenu;
 import com.backendsyndicate.smashclub.admin.model.AdminMenuCategory;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AdminMenuCategoryRepo extends JpaRepository<AdminMenuCategory, Long> {
+public interface AdminMenuCategoryRepo extends JpaRepository<AdminMenuCategory, Integer> {
 
 }
